# api/agent/swagger.yaml
basePath: /v1
definitions:
...
  api.runStep:
    properties:
      error:
        type: string
      finished_at:
        type: string
      id:
        type: string
      input:
        additionalProperties: {}
        type: object
      output:
        additionalProperties: {}
        type: object
      started_at:
        type: string
      status:
        type: string
    type: object
  api.runSummary:
    properties:
      created_at:
        type: string
      error:
        type: string
      finished_at:
        type: string
      id:
        type: string
      inputs:
        additionalProperties: {}
        type: object
      outputs:
        additionalProperties: {}
        type: object
      started_at:
        type: string
      status:
        type: string
      steps:
        items:
          $ref: '#/definitions/api.runStep'
        type: array
      workflow_id:
        type: string
      workflow_version:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    Public HTTP API for the Patchcord agent (vision document, section 10.1). Covers workflow triggering and run observability, connector CRUD and testing, application hosting, and a read-only plugin catalog listing; the rest of the vision document's API surface (actions, full plugin management) is not implemented yet.
    Every route marked with a lock icon below requires "Authorization: Bearer <admin token>" — but only once at least one admin token has been created (`patchcord auth token create`); a fresh agent answers every request unauthenticated, exactly as before this existed (ADR-0036).
  title: Patchcord Agent API
  version: "1"
paths:
...
  /runs:
    get:
      description: Returns every recorded run, most recently created first.
      parameters:
      - description: Restrict to runs of this workflow id
        in: query
        name: workflow_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.runSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List runs
      tags:
      - runs
  /runs/{id}:
    get:
      description: Returns one run's current status, outputs and steps — the counterpart
        to POST /workflows/{id}/run's immediate response, letting a client poll a
        run or fetch its final result once GET /runs/{id}/events has closed.
      parameters:
      - description: Run id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.runSummary'
        "404":
          description: run not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a run
      tags:
      - runs
  /runs/{id}/cancel:
    post:
      description: Marks a run still queued or running as cancelled. Does not interrupt
        a run actively executing within this same process — it only flips its recorded
        status; an in-flight step keeps running until its next persistence checkpoint,
        where it then stops cleanly.
      parameters:
      - description: Run id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.runSummary'
        "404":
          description: run not found
          schema:
            type: string
        "409":
          description: run has already finished
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cancel a run
      tags:
      - runs
...
securityDefinitions:
  BearerAuth:
    description: Admin token, if any has been created (see the top-level description).
      Pass as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"